Updates three new files and added one new file
nameMatching.py, and main.py are updated files. Added new function to nameMatching to output a csv file with separate comments and a matrix that shows whether or not a player is mentioned in said comment. main.py was updated to run new functions.
handCodeCompare.py is a new file which compares machine code (that I created) to a ground truth file in order to calculate recall and precision.